refactor: consolidate `listx` package into `tui` for simplified TUI handling
- Replaced `listx` package references with `tui` across all relevant files.
- Realigned TUI model and run logic for consistency.
- Adjusted import paths and error handling to reflect the package consolidation.
