+++ internal/oci/identity/policy/tui_list.go
@@ -4,13 +4,13 @@ import (
 	"fmt"
 
 	domain "github.com/rozdolsky33/ocloud/internal/domain/identity"
-	"github.com/rozdolsky33/ocloud/internal/tui/listx"
+	"github.com/rozdolsky33/ocloud/internal/tui"
 )
 
 // NewPoliciesListModel builds a TUI list for policies.
-func NewPoliciesListModel(p []domain.Policy) listx.Model {
-	return listx.NewModel("Policies", p, func(p domain.Policy) listx.ResourceItemData {
-		return listx.ResourceItemData{
+func NewPoliciesListModel(p []domain.Policy) tui.Model {
+	return tui.NewModel("Policies", p, func(p domain.Policy) tui.ResourceItemData {
+		return tui.ResourceItemData{
 			ID:          p.ID,
 			Title:       p.Name,
 			Description: fmt.Sprint(p.Description, "  •  ", p.TimeCreated.Format("2006-01-02")),
