+++ internal/services/compute/instance/list.go
@@ -8,7 +8,7 @@ import (
 	"github.com/rozdolsky33/ocloud/internal/app"
 	"github.com/rozdolsky33/ocloud/internal/oci"
 	ociInst "github.com/rozdolsky33/ocloud/internal/oci/compute/instance"
-	"github.com/rozdolsky33/ocloud/internal/tui/listx"
+	"github.com/rozdolsky33/ocloud/internal/tui"
 )
 
 // ListInstances lists instances in a formatted table or JSON format.
@@ -36,9 +36,9 @@ func ListInstances(appCtx *app.ApplicationContext, useJSON bool) error {
 
 	//TUI
 	model := ociInst.NewImageListModel(allInstances)
-	id, err := listx.Run(model)
+	id, err := tui.Run(model)
 	if err != nil {
-		if errors.Is(err, listx.ErrCancelled) {
+		if errors.Is(err, tui.ErrCancelled) {
 			return nil
 		}
 		return fmt.Errorf("selecting instance: %w", err)
